114-whatsapp-business: Create base class and two toucan formats
